requirements.txt
anomalib[full]
cloudinary
flask
gdown
numpy
opencv-python
omegaconf
pillow
python-dotenv
pytorch-lightning==2.4.0
requests
torch==2.4.1
torchvision
gunicorn
runpod